Create another test case for testing all the features in the base class.
